+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 1.1.6 (branch: 1.1.0-stable)
 
+### Resolved Issues
+
+* \#45 When providing database names with invalid characters in them, Moped
+  will now raise an `InvalidDatabaseName` error.
+
 * \#41 `ObjectId.from_time` now only includes the timestamp, no machine or
   process information.